+++ protocols/mdns/src/behaviour/timer.rs
@@ -18,7 +18,7 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-use std::time::{Duration, Instant};
+use instant::{Duration, Instant};
 
 /// Simple wrapper for the different type of timers
 #[derive(Debug)]